chore: move frontend files (from modules.tsx to modules)
