+++ src/modules/CourseMiniCard.tsx
