+++ src/modules/CourseNormalCard.tsx
